small things resolved

+++ src/Core/RequestHandler/Request.php
@@ -45,7 +45,9 @@ class Request extends Response
 
     protected function delete($obj, $get, $post, $input, $error = null, $id = null)
     {
-        $obj->delete();
+        if (!is_null($obj)) {
+            $obj->delete();
+        }
         $this->generateResponse(null, $error);
     }
 }
